requirements.txt
socket
os
zlib
zipfile
threading
re
mysql.connector
pycrypto
pycryptodome